+++ source/utilities/template/vsf_bitmap.c
@@ -39,20 +39,20 @@ extern int_fast8_t __vsf_arch_ffs(uintalu_t);
  *! \param this_ptr    address of the target bitmap
  *! \param biesize  the size of the bitmap in bits
  */
-void __vsf_bitmap_reset(uintalu_t *this_ptr, int_fast16_t bit_size)
+void __vsf_bitmap_reset(uintalu_t *this_ptr, int_fast32_t bit_size)
 {
-    memset(this_ptr, 0, ((uint_fast16_t)bit_size + 7) >> 3);
+    memset(this_ptr, 0, ((uint_fast32_t)bit_size + 7) >> 3);
 }
 
 
-int_fast16_t __vsf_bitmap_ffz(uintalu_t *bitmap_ptr, int_fast16_t bit_size)
+int_fast32_t __vsf_bitmap_ffz(uintalu_t *bitmap_ptr, int_fast32_t bit_size)
 {
-    int_fast16_t word_size =    (bit_size + (int_fast16_t)__optimal_bit_sz - 1) 
-                            /   (int_fast16_t)__optimal_bit_sz;
-    int_fast16_t index = 0, i = 0;
+    int_fast32_t word_size =    (bit_size + (int_fast32_t)__optimal_bit_sz - 1) 
+                            /   (int_fast32_t)__optimal_bit_sz;
+    int_fast32_t index = 0, i = 0;
 
     for (; i < word_size; i++) {
-        int_fast16_t temp = __vsf_arch_ffz(bitmap_ptr[i]);
+        int_fast32_t temp = __vsf_arch_ffz(bitmap_ptr[i]);
         if (temp >= 0) {
             index += temp;
             return index >= bit_size ? -1 : index;
@@ -62,14 +62,14 @@ int_fast16_t __vsf_bitmap_ffz(uintalu_t *bitmap_ptr, int_fast16_t bit_size)
     return -1;
 }
 
-int_fast16_t __vsf_bitmap_ffs(uintalu_t *bitmap_ptr, int_fast16_t bit_size)
+int_fast32_t __vsf_bitmap_ffs(uintalu_t *bitmap_ptr, int_fast32_t bit_size)
 {
-    int_fast16_t word_size =    (bit_size + (int_fast16_t)__optimal_bit_sz - 1) 
-                            /   (int_fast16_t)__optimal_bit_sz;
-    int_fast16_t index = 0, i = 0;
+    int_fast32_t word_size =    (bit_size + (int_fast32_t)__optimal_bit_sz - 1) 
+                            /   (int_fast32_t)__optimal_bit_sz;
+    int_fast32_t index = 0, i = 0;
 
     for (; i < word_size; i++) {
-        int_fast16_t temp = __vsf_arch_ffs(bitmap_ptr[i]);
+        int_fast32_t temp = __vsf_arch_ffs(bitmap_ptr[i]);
         if (temp >= 0) {
             index += temp;
             return index >= bit_size ? -1 : index;
